fix android build
