+++ Plugins/NativeXr/Source/XRHand.h
@@ -4,6 +4,38 @@ namespace Babylon
 {
     namespace
     {
+        constexpr std::array<const char*, 25> HAND_JOINT_NAMES{
+            "wrist",
+
+            "thumb-metacarpal",
+            "thumb-phalanx-proximal",
+            "thumb-phalanx-distal",
+            "thumb-tip",
+
+            "index-finger-metacarpal",
+            "index-finger-phalanx-proximal",
+            "index-finger-phalanx-intermediate",
+            "index-finger-phalanx-distal",
+            "index-finger-tip",
+
+            "middle-finger-metacarpal",
+            "middle-finger-phalanx-proximal",
+            "middle-finger-phalanx-intermediate",
+            "middle-finger-phalanx-distal",
+            "middle-finger-tip",
+
+            "ring-finger-metacarpal",
+            "ring-finger-phalanx-proximal",
+            "ring-finger-phalanx-intermediate",
+            "ring-finger-phalanx-distal",
+            "ring-finger-tip",
+
+            "pinky-finger-metacarpal",
+            "pinky-finger-phalanx-proximal",
+            "pinky-finger-phalanx-intermediate",
+            "pinky-finger-phalanx-distal",
+            "pinky-finger-tip" };
+
         class XRHand : public Napi::ObjectWrap<XRHand>
         {
             static constexpr auto JS_CLASS_NAME = "XRHand";
